+++ src/api/events/ui.ts
@@ -1,6 +1,6 @@
+import type { JsonValue } from "../../common/json";
 import type { ActionEventMessage } from "./action";
 import type { DeviceIdentifier } from "./device";
-import type { PayloadObject } from "./index";
 
 /**
  * Occurs when the property inspector associated with the action becomes visible, i.e. the user selected an action in the Stream Deck application. Also see {@link PropertyInspectorDidDisappear}.
@@ -15,7 +15,7 @@ export type PropertyInspectorDidDisappear = ActionEventMessage<"propertyInspecto
 /**
  * Message sent between the plugin and it's respective UI.
  */
-type PluginMessage<TEvent extends string, TPayload extends PayloadObject<TPayload>> = Omit<ActionEventMessage<TEvent>, keyof DeviceIdentifier> & {
+type PluginMessage<TEvent extends string, TPayload extends JsonValue> = Omit<ActionEventMessage<TEvent>, keyof DeviceIdentifier> & {
 	/**
 	 * Payload sent between the plugin and it's UI.
 	 */
@@ -25,9 +25,9 @@ type PluginMessage<TEvent extends string, TPayload extends PayloadObject<TPayloa
 /**
  * Occurs when a payload was received from the UI.
  */
-export type DidReceivePropertyInspectorMessage<TPayload extends PayloadObject<TPayload>> = PluginMessage<"sendToPlugin", TPayload>;
+export type DidReceivePropertyInspectorMessage<TPayload extends JsonValue> = PluginMessage<"sendToPlugin", TPayload>;
 
 /**
  * Occurs when a message was received from the plugin.
  */
-export type DidReceivePluginMessage<TPayload extends PayloadObject<TPayload>> = PluginMessage<"sendToPropertyInspector", TPayload>;
+export type DidReceivePluginMessage<TPayload extends JsonValue> = PluginMessage<"sendToPropertyInspector", TPayload>;