+++ src/common/json.ts
@@ -1,7 +1,9 @@
 /**
- * Credit to Maksim Zemskov
- * https://hackernoon.com/mastering-type-safe-json-serialization-in-typescript
+ * JSON object.
  */
+export type JsonObject = {
+	[key: string]: JsonValue;
+};
 
 /**
  * JSON primitive value.
@@ -11,24 +13,4 @@ export type JsonPrimitive = boolean | number | string | null | undefined;
 /**
  * JSON value.
  */
-export type JsonValue =
-	| JsonPrimitive
-	| JsonValue[]
-	| {
-			[key: string]: JsonValue;
-	  };
-
-/**
- * JSON compatible value.
- */
-export type JsonCompatible<T> = unknown extends T
-	? never
-	: {
-			[P in keyof T]: T[P] extends JsonValue ? T[P] : T[P] extends NotAssignableToJson ? never : JsonCompatible<T[P]>;
-	  };
-
-/**
- * Values that aren't assignable to JSON.
- */
-// eslint-disable-next-line @typescript-eslint/ban-types
-type NotAssignableToJson = Function | bigint | symbol;
+export type JsonValue = JsonObject | JsonPrimitive | JsonValue[];