+++ src/plugin/events/ui-message-event.ts
@@ -1,13 +1,12 @@
-import type { DidReceivePropertyInspectorMessage, PayloadObject } from "../../api";
+import type { DidReceivePropertyInspectorMessage } from "../../api";
 import { Event } from "../../common/events";
+import type { JsonObject, JsonValue } from "../../common/json";
 import type { Action } from "../actions/action";
 
 /**
  * Provides information for an event triggered by a message being sent to the plugin, from the property inspector.
  */
-export class DidReceivePropertyInspectorMessageEvent<TPayload extends PayloadObject<TPayload>, TSettings extends PayloadObject<TSettings>> extends Event<
-	DidReceivePropertyInspectorMessage<TPayload>
-> {
+export class DidReceivePropertyInspectorMessageEvent<TPayload extends JsonValue, TSettings extends JsonObject> extends Event<DidReceivePropertyInspectorMessage<TPayload>> {
 	/**
 	 * Payload sent from the property inspector.
 	 */