+++ src/ui/action.ts
@@ -1,10 +1,10 @@
-import type { PayloadObject } from "../api";
+import type { JsonObject } from "../common/json";
 
 /**
  * Provides a contextualized instance of an {@link Action}, allowing for direct communication with the Stream Deck.
  * @template T The type of settings associated with the action.
  */
-export type Action<T extends PayloadObject<T>> = {
+export type Action<T extends JsonObject> = {
 	/**
 	 * Unique identifier of the instance of the action; this can be used to update the action on the Stream Deck, e.g. its title, settings, etc.
 	 */
@@ -20,7 +20,7 @@ export type Action<T extends PayloadObject<T>> = {
 	 * @template U The type of settings associated with the action.
 	 * @returns Promise containing the action instance's settings.
 	 */
-	getSettings<U extends PayloadObject<U> = T>(): Promise<U>;
+	getSettings<U extends JsonObject = T>(): Promise<U>;
 
 	/**
 	 * Sets the {@link settings} associated with this action instance. Use in conjunction with {@link Action.getSettings}.