+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ### 🐞 Fix
 
 -   `Coordinates` type could erroneously have a non-number type for `row`.
+-   Fix support for allowed types within payloads.
 
 ### ♻️ Update
 
@@ -32,6 +33,7 @@
 
 ### ➡️ Migration
 
+-   `PayloadObject<T>` replaced with `JsonObject`.
 -   `streamDeck.ui.onSendToPlugin` renamed to `streamDeck.ui.onDidReceivePropertyInspectorMessage`.
 -   `SendToPluginEvent<T>` renamed to `DidReceivePropertyInspectorPayloadEvent<T>`.
 -   JSON schemas have been relocated to a dedicated schemas package, [`@elgato/schemas`](https://github.com/elgatosf/schemas).